+++ Anubis.LC.LaserControlPlugin/Components/LaserPointerTargetLandmine.cs
@@ -1,5 +1,6 @@
 ﻿using Anubis.LC.LaserControlPlugin.Extensions;
 using Anubis.LC.LaserControlPlugin.Helpers;
+using Anubis.LC.LaserControlPlugin.ModNetwork;
 using UnityEngine;
 
 namespace Anubis.LC.LaserControlPlugin.Components
@@ -18,9 +19,7 @@ namespace Anubis.LC.LaserControlPlugin.Components
 
         public override void Warmup(Transform origin, FlashlightItem laserPointer)
         {
-            if (!LethalConfigHelper.IsPointerCanDetonateLandmines.Value) return;
-            if (!target) return;
-            if (target.hasExploded) return;
+            if (!Networking.Instance.GetConfigItemValueOfPlayer(nameof(LethalConfigHelper.IsPointerCanDetonateLandmines)) || !target || target.hasExploded) return;
 
             if (!triggered && HasCollision(origin.position, origin.forward, transform.position + offset, radius)
                  && !Physics.Linecast(origin.position, hitPoint, 1051400, QueryTriggerInteraction.Ignore))