+++ Anubis.LC.LaserControlPlugin/Components/LaserPointerTargetTurret.cs
@@ -22,8 +22,7 @@ namespace Anubis.LC.LaserControlPlugin.Components
 
         public override void Warmup(Transform origin, FlashlightItem laserPointer)
         {
-            if (!LethalConfigHelper.IsPointerCanTurnOnAndOffTurrets.Value) return;
-            if (!target) return;
+            if (!Networking.Instance.GetConfigItemValueOfPlayer(nameof(LethalConfigHelper.IsPointerCanTurnOnAndOffTurrets)) || !target) return;
 
             if (!triggered && HasCollision(origin.position, origin.forward, transform.position + offset, radius)
                  && !Physics.Linecast(origin.position, hitPoint, 1051400, QueryTriggerInteraction.Ignore))